application/requirements.txt
certifi==2018.1.18
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
decorator==4.0.11
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.0.2
django-common-helpers==0.9.1
django-crispy-forms==1.7.0
django-cron==0.5.0
django-filter==1.1.0
django-widget-tweaks==1.4.1
djangorestframework==3.8.2
gunicorn==19.7.1
httplib2==0.11.3
idna==2.6
imageio==2.1.2
InstagramAPI==1.0.2
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.11
MarkupSafe==1.0
moviepy==0.2.3.2
numpy==1.14.3
oauthlib==2.0.6
Pillow==5.0.0
psycopg2==2.7.4
psycopg2-binary==2.7.4
PyJWT==1.6.0
python-instagram==1.3.2
python3-openid==3.1.0
pytz==2018.3
requests==2.11.1
requests-oauthlib==0.8.0
requests-toolbelt==0.7.0
simplejson==3.15.0
six==1.11.0
social-auth-app-django==2.1.0
social-auth-core==1.7.0
tqdm==4.11.2
uritemplate==3.0.0
urllib3==1.22
